+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com), and this project adheres to [Semantic Versioning](https://semver.org).
 
+## [0.11.1] - 2021-03-21
+
+### Fixed
+
+- Alternate links in the head.
+
 ## [0.11.0] - 2021-03-21
 
 ### Added
@@ -274,7 +280,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 - Support for internationalizing pages.
 - Both light and dark themes for the site.
 
-[unreleased]: https://github.com/visiosto/visiosto.fi/compare/v0.11.0...HEAD
+[unreleased]: https://github.com/visiosto/visiosto.fi/compare/v0.11.1...HEAD
+[0.11.1]: https://github.com/visiosto/visiosto.fi/compare/v0.11.0...v0.11.1
 [0.11.0]: https://github.com/visiosto/visiosto.fi/compare/v0.10.3...v0.11.0
 [0.10.3]: https://github.com/visiosto/visiosto.fi/compare/v0.10.2...v0.10.3
 [0.10.2]: https://github.com/visiosto/visiosto.fi/compare/v0.10.1...v0.10.2